Terms Query: Remove some redundant attributes
