+++ packages/block-library/src/term-template/index.php
@@ -32,12 +32,10 @@ function render_block_core_term_template( $attributes, $content, $block ) {
 
 	$query_args = array(
 		'taxonomy'   => $query['taxonomy'] ?? 'category',
-		'number'     => $query['perPage'] ?? 10,
+		'number'     => $query['maxTerms'] ?? 10,
 		'order'      => $query['order'] ?? 'asc',
 		'orderby'    => $query['orderBy'] ?? 'name',
 		'hide_empty' => $query['hideEmpty'] ?? true,
-		'include'    => $query['include'] ?? array(),
-		'exclude'    => $query['exclude'] ?? array(),
 	);
 
 	// We set parent to 0 only if we show all terms as hierarchical or we show top-level terms.
