+++ packages/block-library/src/terms-query/block.json
@@ -9,20 +9,14 @@
 	"keywords": [ "terms", "taxonomy", "categories", "tags", "list" ],
 	"textdomain": "default",
 	"attributes": {
-		"termQueryId": {
-			"type": "number"
-		},
 		"termQuery": {
 			"type": "object",
 			"default": {
-				"perPage": 10,
-				"pages": 0,
+				"maxTerms": 10,
 				"taxonomy": "category",
 				"order": "asc",
 				"orderBy": "name",
 				"hideEmpty": true,
-				"include": [],
-				"exclude": [],
 				"parent": false,
 				"hierarchical": false
 			}
@@ -35,13 +29,9 @@
 		"tagName": {
 			"type": "string",
 			"default": "div"
-		},
-		"namespace": {
-			"type": "string"
 		}
 	},
 	"providesContext": {
-		"termQueryId": "termQueryId",
 		"termQuery": "termQuery",
 		"termsToShow": "termsToShow"
 	},
