+++ test/integration/fixtures/blocks/core__terms-query.html
@@ -1,3 +1,3 @@
-<!-- wp:terms-query {"termQueryId":0} -->
+<!-- wp:terms-query -->
 <div class="wp-block-terms-query"></div>
 <!-- /wp:terms-query -->
