+++ test/integration/fixtures/blocks/core__terms-query.json
@@ -5,7 +5,7 @@
 		"attributes": {
 			"originalName": "core/terms-query",
 			"originalUndelimitedContent": "<div class=\"wp-block-terms-query\"></div>",
-			"originalContent": "<!-- wp:terms-query {\"termQueryId\":0} -->\n<div class=\"wp-block-terms-query\"></div>\n<!-- /wp:terms-query -->"
+			"originalContent": "<!-- wp:terms-query -->\n<div class=\"wp-block-terms-query\"></div>\n<!-- /wp:terms-query -->"
 		},
 		"innerBlocks": []
 	}
